Remove server-starting script
The server can be started by running `finales server start` and
providing any necessary options, so the script in the `server`
folder is no longer necessary (some of the remaining testing
endpoints that were still there are now migrated to the CLI).
