+++ src/FINALES2/cli/server.py
@@ -1,7 +1,9 @@
 import click
 import uvicorn
-from fastapi import FastAPI
+from fastapi import Depends, FastAPI
 
+from FINALES2.schemas import User
+from FINALES2.server import config
 from FINALES2.server.operations import operations_router
 from FINALES2.userManagement import userManager
 
@@ -39,4 +41,16 @@ def server_start(ip, port):
     )
     app.include_router(router=userManager.userRouter)
     app.include_router(router=operations_router)
+
+    @app.get("/")
+    def Hello():
+        """Remainder from the first script to start the server."""
+        return "Hello! This is FINALES2."
+
+    @app.get("/test")
+    def test(token: User = Depends(userManager.getActiveUser)):
+        """Remainder from the first script to start the server."""
+        print(config.userDB)
+        return token
+
     uvicorn.run(app=app, host=ip, port=port)
